Move metrics fixtures to the fixtures file
