+++ src/decisionengine/framework/engine/tests/fixtures.py
@@ -70,6 +70,8 @@ class DETestWorker(threading.Thread):
         self.global_config["server_address"] = self.server_address
         self.global_config["dataspace"]["datasource"] = datasource
         self.global_config["no_webserver"] = True
+        self.global_config["webserver"] = {}
+        self.global_config["webserver"]["port"] = get_random_port()
 
         self.de_server = _create_de_server(self.global_config, self.channel_config_loader)
         self.stdout_at_setup = None
